requirements.txt
-f https://download.pytorch.org/whl/cpu
torch==2.1.2
torchvision==0.16.2
pillow==9.4.0